// EscalafonFrontend/src/app/interfaces/relacion-individual-colectiva.ts
export interface RelacionIndividualColectiva {
    /**
     * @type { number } - iRelaIndColecId
     */
    iRelaIndColecId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { Date } - dtRelaIndColecFechaEmision
     */
    dtRelaIndColecFechaEmision?: Date;

    /**
     * @type { string } - cRelaIndColecAnotaciones
     */
    cRelaIndColecAnotaciones?: string;

}